Debugged ability HUD for remote clients
The new modular ability HUD now works for listen servers and for clients. This required the HUD system to be refactored to utilize pure listening, rather than binding to ability specs. This is because a lot of data in abilities' actor info has not been replicated by the time we try to initialize their HUD.

This new system requires every ability to define its own widget (which can just be subclassed for W_AbilityWithCooldown, W_UltimateAbility, or whatever base ability widget is appropriate), and to have a unique gameplay tag (e.g. Ability.Identifier.Knight.Joust) to identify which ability messages are relevant to it.

I dislike the extra work required to set up ability widgets with this system, and the limited ability-related data we have access to in ability widgets. But considering this is the most viable method for a consistent, reliable HUD system, there isn't much of a choice. The silver lining is a slight increase in flexibility, as we no longer require every ability widget to be of type UAbilityWidgetBase.