Created the AbilitySystemWidgetBase widget
This is a new user widget that automatically binds to its owning player's ASC to provide easy access in BP. This will be used for things like attribute displays (e.g. the "Battle Rage" meter), as it can easily bind to attribute sets and delegates as soon as they're available.

Reparented the AbilityWidgetBase class to this new class to give it simple access to its owner's ASC, which no longer has to be passed around manually to each routed event.