Implemented tag-based team lives
This method uses team tags to track lives. When one player dies, their team's "Lives" tag stack is decremented. When players die, they can only respawn if their team has remaining "Lives" tags.

I'm not a HUGE fan of this method, because it doesn't allow for prediction, and feels too exposed for something as low-level as lives. I would prefer to implement a "Lives" attribute set, but that would also introduce a lot of unnecessary complications that are solved by how simple this tag system is.

Since these tags aren't predicted but respawning IS predicted (since it needs to execute a lot of client-side logic), we need a hack to guess whether we'll be able to respawn or not (we check for 0 "Lives" tags on the server, but we check for 1 "Lives" tag on the client, since it won't be replicated yet). I'm currently working to find a way to make the initial respawn check in the auto respawn ability (OnDeathStarted) server-side, while still allowing the rest of the ability to execute client-side logic.