Removed "Lives" attribute set
Team lives will not be managed via an attribute set. Doing this obscures what should be a very simple system by adding an enormous amount of boilerplate code, for what should really be a system exposed to designers. This would also make updating and reading team lives more difficult, and reliant on an abstracted external system (messaging) rather than declaratively managing lives in relevant classes (e.g. the death ability).

At the end of the day, "team lives" are more like a team score than a player attribute. Therefore lives will instead be tracked using team tags—a system for which this exact concept was designed. Explicitly handling something so relevant to the game mode is friendlier for designers, more maintainable, and avoids relying on an entire C++ class for such a high-level mechanic—which is the entire point of our modular game mode framework.