Implemented server-side auto-respawn
This is an alternate implementation of the current auto-respawn ability, which uses explicitly server-side logic, as opposed to predicted client-side logic. This prevents clients from having to guess the result of the respawn ability (e.g. whether or not they can respawn, as this is dictated by the server) while still allowing them to implement client-side logic, like triggering widget effects or switching to spectator mode.

I find the current respawn implementation to be significantly more elegant than this new version, and it still works perfectly well. So I'm not switching to the new implementation just yet; I want to see if we can "get away with" using this less safe but more elegant version, or if it will become such a hassle that we have to switch.