Created the WidgetBase gameplay cue
This is a new gameplay cue that can be used to spawn a user widget while the cue is active. This essentially hijacks the gameplay cue system to be used like a "spawn widget" RPC, while also allowing us to add and remove widgets from within GAS features (e.g. widgets that are visible while a gameplay effect is active).