Version 0.3.6
This version implements the "game phase subsystem." This is a new server-side world subsystem that manages the "phases" of the game; e.g. character selection, countdown, playing, overtime, post-game, etc. To test this subsystem, "count-down" and "playing" phases have been implemented, including a count-down animation when the game starts, and various game mode-specific HUD information. A new gameplay debugger has also been added to aid in debugging game phases.

This version also implements "lives," which are shared by all players of a team, and implemented as simple gameplay tags in the team info actor. When a player dies, their team loses a life; when a team runs out of lives, they lose the game.

This version also includes minor improvements to various systems, such as death and respawn mechanics.