Updated how knockback forces are calculated and applied
Previously, knockback was applied using an impulse on characters' movement components. This has been updated to use the LaunchCharacter method for applying knockback to characters, as physics should not be taken into account.

Forced upward velocity has also been updated. Previously, a significant upwards force was always applied, when requested, when using the AddKnockbackToTargetFromLocation method, but not used when applying knockback with a direction vector (AddKnockbackToTargetInDirection). This has been updated such that an upward velocity is now applied automatically, regardless of the knockback method, but only when the target actor is on the ground. This allows for more simple, intuitive, and predictable knockback results, while still avoiding sliding and friction when applying knockback to grounded actors.