Added perspective support to equipment animations
Equipment actors can now play perspective-specific animations. These can be triggered by the "PlayEquipmentAnimation" animation notify, which asks to specify which perspective to trigger the animation for.

If you want to use the same animation for both perspectives (usually the case), you can just trigger a notify in both character animations that uses the same animation.