Implemented projectile attachment to Knight's "Spear" projectile
The projectile used for the Knight's "Spear" ability now attaches to characters it hits.

The projectile now also uses an additional hit detection method, in the form of a new collision component that only detects hits with valid targets (actors with an ASC). This lets us give the spear a more generous hitbox, while preventing it from hitting the environment in unrealistic ways.