Added socket-based filtering to "HideEquipment" anim notify
The "HideEquipment" animation notify can now be used to hide equipment attached to specific bones/sockets. An editor-exposed parameter has been added that allows animators to specify an bone or socket name. If this name is set, only equipment actors attached to bones or sockets of that name will be hidden.

This was added to support single-armed animations for the Knight character, who uses one equipment actor in each hand. Now, we can make ability animations that only use one hand, hiding the equipment actor in that hand, and keeping the other equipment actor visible.