Implemented "TriggerAbilityTargeting" anim notify state
Created a new anim notify state, which will be used to determine the targeting duration of abilities with event-based targeting. This will be useful, for example, for determining when a melee attack can hit targets during its animation.

Creating a notify state for this, rather than using the existing "TriggerGameplayEvent" notify, allows us to avoid having to create two event tags for every ability that wants to control when its targeting starts and ends with events (i.e. "StartTargeting" and "EndTargeting").

This notify will primarily be used for the base melee attack ability, as it does require some validation when receiving the events, to ensure that the correct ability's targeting is triggered.