Created outline for base "Continuous Trace" target actor
This will be a custom reusable GA target actor that will perform a continuous trace throughout it's activation. It will primarily be used to provide continuous hit detection for melee attacks.

This is a rewrite of the engine's default trace actor, as it does not support multiple hits and is not well-suited to being reused (via the "Wait Target Data with Reusable Actor" task).