Implemented basic continuous tracing logic for target actors
Implemented the basic logic for continuous tracing with the "ContinuousTrace" target actor. Target actors of this class can now continuously perform line traces to detect targets and send them to the server for validation and processing.

Currently, the "Cone" continuous trace target actor performs line traces, which was used to feature testing. It will be updated to perform the correct cone-shaped traces.