Removed anim instance caching from actor info structure
Animation instances should never be replicated, and the "Actor Info" structure should not be conditionally serialized to account for this. Instead, the cached animation instance variables in the actor info structure have been removed. Any potential use of these variables should instead utilize the "GetAnimInstance" functions, which retrieve the animation instance from the cached mesh components, which SHOULD be serialized.