Moved ability event triggers to third-person animations
Some animation-based ability event triggers were placed in their ability's first-person animation. This would be mistake, because the first-person mesh is not guaranteed to always be in sync with the network, while the third-person mesh is. These events have been moved to the third-person mesh and a note has been added in the event trigger notify for future reference.