Added hit filtering for better direction approximations to cone target actor
Previously, the cone target actor would take the first hit against an actor, and throw out any subsequent hits on that same actor to save processing time. But this often led to situations where the hit was far away from where the player was actually aiming (e.g. the edge of the cone hitting an enemy's capsule). This had no gameplay consequences, but could cause VFX to spawn far away from where they would be expected: where the player's aiming.

To fix this, repeated hits on actors are no longer thrown out. Instead, they'll be fully processed and compared to the existing hit on the same actor. If the new hit is closer to the player's camera than the existing one (dot is closer to 1.0), then the new hit will be used instead of the old one.

Now, when a player hits an actor using the cone trace target actor, the processed hit will be the one closest to the player's camera, ensuring VFX always spawn as close to the expected location as possible.