Implemented Knight's "Inspire" ability
This is a passive ability that grants the Knight stacks of 8% damage boost when hitting an enemy with an ability (Shortsword, Shield Bash, Warhammer, or Spear).

Currently, the ability is also procced by hitting interactable actors (e.g. explosive barrels) with abilities. This should be fixed in the next commit.