Implemented basic "Widget" GE component
This is a new GE component that can push a user widget to the local player when a gameplay effect is added, and can communicate updates to and removal of that effect to the widget.

This feature is fully functional, with the exception of broadcasting stack updates to clients. This will be fixed in a future commit.