Implemented Knight's "Inspire" indicator widget
Implemented functionality for the indicator widget for the Knight's "Inspire" passive ability.

This works by registering an indicator widget when the ability is granted, and broadcasting a message to that widget whenever the effect is applied or stacked. Because stacks are not replicated to clients, we have to broadcast these messages on the server.