Updated base radial material to support partial transparency for non-segmented materials
This allows us to create partial rings that can be filled relative to their portion of a circle. A semi-circle that can be filled without having the rest of the ring visible. This is used for the reticle indicator for the Knight's passive "Inspire" ability.

Refactored all existing radial materials to use the newer material, without changes to their appearance. Previous versions have been backed up as a duplicate asset labeled "Old" until new materials can be validated to ensure their appearance has not been changed.