Minor changes to damage interaction filtering
The damage interaction check performed in the damage execution was largely unnecessary, because we usually filter out unwanted targets in our target actor, since we typically want to filter out other effects besides damage (e.g. knockback) depending on the interaction type. Doing it again in the damage execution served practically no purpose.

Instead of removing it, the damage interaction check has been updated to  take into account the dynamic tags of the damage effect, to allow designers to control how the damage effect is applied. For example, the "CanDamageSelf" tag can now be added to the damage effect to allow the effect to damage its instigating ASC.

This change was made to give this check a new purpose: filtering damage separately from other effects. For example, if we want to make a rocket-jump that applies knockback to its owner, but not damage, we can do that without having to perform additional filtering in the attack itself. Instead, we would allow the rocket's explosion to detect its owner, so knockback can be applied, but we wouldn't give the damage effect a "CanDamageSelf" tag, so the knockback is still processed, but the damage execution is thrown out.

Self/team damage is disallowed by default, because we very rarely want to damage ourself or our teammates. So when we do, we want to be very explicit and deliberate about it. In other words, we're not allowing the user to specify the rules of their damage application, but rather giving them an option to override the default behavior, if desired, using effect tags.