Changed Knight's "Spear" ability spawn transform logic
The Knight's "Spear" ability previously played an animation, waited for a certain point in that animation, then spawned the projectile in the character's hand, rotated such that it was directed towards whatever the player was aiming at.

This method looked great, but often felt terrible. The animation caused an annoying delay between the ability's activation and the projectile actually firing, while the projectile's path from the character's hand to whatever the player was looking at (which may not always be what they're aiming for) often led to the projectile not going where the player wanted it to, and the player having to aim in an unintuitive way for the projectile to land where they wanted it to, having to account for the projectile's drop, angle from the hand to the camera, and spawning delay.

To remedy these issues, the projectile now spawns instantly at the player's camera, rotated in the direction the player is aiming. These doesn't look nearly as clean (though may be improved when the animations are adjusted), but feels significantly better, more consistent, and more predictable for the ability's user.