Added new network emulation profiles for latency testing
The lowest amount of packet loss that Unreal allows emulation for is 1%, which is still enormous for any realistic testing. These new network profiles use latency without packet loss for when we want to test gameplay experiences with high latency, without worrying about missing communication between the client and server.