Added a camera mode override to developer settings
Annoyingly, the rewind debugger doesn't work with meshes that are invisible during recording (the third-person mesh in a first-person game), so if we want to debug our animations with it, we need to use a camera mode where the third-person mesh is visible. To make this easier, I've added a developer option to override the default camera mode used during PIE, so we don't have to bother with temporarily modifying the pawn data we're using.