Modified aim offsets in base third-person animation blueprint
Moved aim offsets to be applied before montages are blended in. This will allow the aim offsets to act more like base animations, on top of which additive montages can be played. For example, the Knight's "Shortsword" attack can now be played in the direction of the player's aim, rather than being played normally before being adjusted for pitch.