Removed spine yaw from global base animations
I'm not sure why, but our global base animations previously offset the yaw of the character's upper body without animating it. This meant that the base of every character animation had their yaw offset by X degrees. This has been removed from our base animations, because these are supposed to serve as the base for other animations, and shouldn't determine the position of bones that aren't being additively animated on their own.