Removed "Configure" method from collision detector target actor
Previously, target actors had to be spawned, configured, and attached to characters. Now, the configurable parameters are exposed on spawn, so they don't have to set manually in a separate call. A new parameter, bAttachToCharacter, also removes the need to manually attach the actor on initialization or whenever whenever the avatar changes.