Updated collapsible bridge actors
Collapsible bridge actors are now non-replicated as autonomous proxies, like normal level actors. This is because the collision of these actors should behave like other level actors, and all dynamic changes that occur to them are completely deterministic and should be performed locally.

Also fixed the spelling of their name...