Updated how "Inspire" ability is triggered
The "Inspire" ability is still procced via a gameplay effect, but it no longer uses network messages to notify clients of the application of the gameplay effect. Instead, the effect adds a gameplay tag with each stack, which triggers a "Wait for Gameplay Tag Count Changed" listener in the "Inspire" widget, which then retrieves the active gameplay effect to get its current stack count and duration.

This system uses significantly less code and overhead scripts, and no longer makes chaining network calls for notifying clients.