Fixed bug preventing Common UI from binding universal actions
Because our gameplay input is quite complex and needs to be flexible (e.g. configurable by players), we've been using Enhanced Input. But setting up UI input with Enhanced Input would be a nightmare, because we'd need to manage a collection of new input actions, as well as mapping contexts to bind those actions. Since our UI input doesn't need to be sophisticated at all (we don't need to configure it, we don't need any complex functionality, etc.), we use Common UI's input framework for it instead.

This creates a problem, because Enhanced Input and Common UI input aren't meant to be mixed. We've made some minor, but very important, changes to the Common UI code to prevent the plugin from breaking when we try to bind Common UI input while Enhanced Input is enabled. No functional changes have been made to the game code, but an important comment explaining the technicalities of this issue has been added.