Completed implementation of base common action widget
This widget can be used to display the appropriate brush to represent a certain input action (e.g. "L2" for an ability activated by the left trigger), or to retrieve the display name of the keybinding, if no predefined brush has been set (e.g. "LShift" for an ability activated by the left shift key).

This new widget is used to display the key bindings of abilities in the game HUD. The widget's code still needs to be cleaned up and reformatted.