Minor refactoring of base ability widget
Reformatted and cleaned up code.

Also switched from using an identifying gameplay tag to using a class pointer to bind the widget's ability. This is a fairly arbitrary choice, since we only use this variable when the widget is created, and after that, we can just use the ability itself; we also should never have multiple specs of the same ability class, nor multiple specs with the same identifying tag. But I'm choosing to use the class itself, rather than a tag, just to make the setup process a little more intuitive. This also lets us leverage the FindAbilitySpecFromClass function, which is faster than searching every tag on every ability.

All gameplay abilities still have identifying tags, which are used for other queries.