Updated Xbox platform input key textures
The textures for the input keys on the Xbox gamepad have been updated to fit the size and color scheme used in this project.

Also updated the base "Ability with Cooldown" widget to dynamically scale icons to be as large as they can, rather than always forcing them into a 64x square.