Added "Invulnerable" status to ability target filtering
Target actors, abilities, and projectiles now automatically filter out actors with the "Invulnerable" status effect.

Currently, this status effect makes the actor immune to abilities, but still allows them to be collide with projectiles and melee attacks, meaning they aren't "phased out," just immune to abilities. I may decide to change this in the future if I want the "Invulnerable" status to phase out the actor entirely, so projectiles and weapons completely ignore them, instead of hitting but not affecting them.