Fixed bug causing duplicate ability widgets on listen servers
OnGiveAbility is called on the server and the client when an ability is granted. Previously we were not doing any kind of check to ensure ability widgets, which are added during by function, were only added on the ability's owning client, causing every client's ability widgets to be added on the server each time they were granted.

To fix this, we've added a small check before adding widgets, to ensure we're only doing so locally. When this function is called, the ability's actor info may not be replicated to clients yet, so we have to use the player state given by the function's ActorInfo parameter to check whether we're local or not.