Created first-person directional movement aim offsets for Viking
Also added a new setting to the base first-person animation blueprint: "Use Directional Speed for Walking BS". If true, directional, signed velocity will be used as parameters for the "Idle/Walk" blend space. If false, the character's absolute velocity magnitude will be used instead. The former is how this blueprint previously worked, to support directional walking animations. The latter is a new option for characters that want to use a single non-directional walking animation instead.