Switched to using per-effect gameplay effect classes
Previously, we were using a reusable effect class, and setting its parameters during runtime with SetByCaller modifiers. This was becoming a nightmare to work with, as we had to set not only damage magnitude, but effect context data, and knockback information; it was a huge amount of boilerplate code we had to re-write every time we just wanted to apply damage. Using a per-effect class structure instead makes scripting effects a lot easier.