Adjusted how trace hit location override offsets are applied for bone targets
Previously, the offset applied to overridden trace hit locations when using bone locations (i.e. when the player was not aiming at a character mesh) was calculated using the original trace's hit normal. This resulted in offsets being applied in undesirable ways. For example, if the player was aiming upwards, above a target, and hit them, the hit location would be offset downwards, since it was applied in the direction opposite to the player's original aim normal.

The way in which offsets are calculated has been updated to instead use the normal from the ADJUSTED hit location to the hit's origin. In other words, rather than offsetting the hit location opposite to the player's aim, we are now offsetting the hit location towards the player's camera (or wherever the trace originated).