Added the "NoUltimateCharge" GE tag
This is a new gameplay tag that can be added to gameplay effects. If damage or healing effects with this tag are applied, ultimate charge will not be granted to the instigator. This is useful for preventing certain gameplay effects from granting ultimate charge.

We do this per-effect instead of per-ability (e.g. having an "Ability.Behavior.NoUltimateCharge" tag) because we rarely pass instigating abilities into effect contexts, and we may want to do this for effects applied WITHOUT abilities (e.g. environmental damage).