Began implementing "Radial Ground Detection" collision utility
This will eventually perform a cylindrical overlap and return overlapped actors that are on the ground, within a certain threshold. This will be used for the Viking's "Earthbreaker" ability, which launches nearby players upwards when the character lands on the ground.

Currently, this utility performs a cylindrical overlap, but does not yet filter for grounded actors.

Also created a new math library for additional utilities.