Removed damage value flooring for periodic damage effects
Normally, damage is floored, as to only apply it in whole numbers (since it's only displayed to players in whole numbers).  If damage is applied periodically, however, the applied damage values will not be floored, to allow it to be applied in smaller increments (e.g. 0.5 damage per tick, at 5 ticks per second). Note that the total amount of damage applied should still be a round number.