Added zero-health-based filtering to damage execution
Damage executions will now throw out any damage applied to targets who don't have any health (health attribute <= 0).

It is still largely the responsibility of gameplay abilities to ignore dead and dying actors (which is how most of these situations are avoided, as dead actors have no hitbox). This check is simply a last resort to ensure damage is never applied to a dead target (especially in the case of asynchronous damage, like DoT).

Also added the "State.Invulnerable" tag to the list of tags targets cannot have in the base damage effect classes. This prevents damage effects from being applied to invulnerable targets. This is not done in the damage execution itself because there are still some special cases where we want to apply damage to invulnerable targets (e.g. if they fall off the map).