Created new "Dash" version of Viking's "Battle Axe" ability
This is a new method of primary attacks that I'm experimenting with, which move the player towards their target as they hit them, like you'd see in a third-person RPG or PvP game. I'm not sure whether I like it more yet.

Added new methods to the collision and ability libraries to create this ability.