Implemented "Is Ability Active" utilities in the ASC
These utilities can be used in C++ or blueprints to check if a given ASC currently has an ability matching a certain query active.

This is used in the Viking's "Earthbreaker" ability to avoid proccing the ability if the Viking's "Earthshatter" ability is active, which triggers a more powerful version instead.